requirements.txt
requests
discord.py
python-dotenv
replicate